Move gitclient to contract
