+++ core/caching.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/huangsam/hotspot/internal/contract"
-	"github.com/huangsam/hotspot/internal/gitclient"
 	"github.com/huangsam/hotspot/internal/iocache"
 	"github.com/huangsam/hotspot/schema"
 )
@@ -17,7 +16,7 @@ import (
 const currentCacheVersion = 1
 
 // cachedAggregateActivity - Simplified and validated using DB columns
-func cachedAggregateActivity(ctx context.Context, cfg *contract.Config, client gitclient.GitClient, mgr iocache.CacheManager) (*schema.AggregateOutput, error) {
+func cachedAggregateActivity(ctx context.Context, cfg *contract.Config, client contract.GitClient, mgr iocache.CacheManager) (*schema.AggregateOutput, error) {
 	activity := mgr.GetActivityStore()
 	if activity == nil {
 		// Fallback to direct computation
@@ -57,7 +56,7 @@ func checkCacheHit(activity iocache.CacheStore, key string) *schema.AggregateOut
 }
 
 // computeAndStore computes the result and stores it in cache
-func computeAndStore(ctx context.Context, cfg *contract.Config, client gitclient.GitClient, activity iocache.CacheStore, key string) (*schema.AggregateOutput, error) {
+func computeAndStore(ctx context.Context, cfg *contract.Config, client contract.GitClient, activity iocache.CacheStore, key string) (*schema.AggregateOutput, error) {
 	result, err := aggregateActivity(ctx, cfg, client)
 	if err != nil {
 		return nil, err
@@ -72,7 +71,7 @@ func computeAndStore(ctx context.Context, cfg *contract.Config, client gitclient
 }
 
 // generateCacheKey creates a unique key based on analysis parameters
-func generateCacheKey(ctx context.Context, cfg *contract.Config, client gitclient.GitClient) string {
+func generateCacheKey(ctx context.Context, cfg *contract.Config, client contract.GitClient) string {
 	// Use canonical helpers from internal.Config to ensure consistent time granularity
 	startHour := cfg.GetAnalysisStartTime()
 	endHour := cfg.GetAnalysisEndTime()
